EventStreamWrapper: add one "overall lifecycle" text fixture
Its purpose is to check that, even in the face of a restart
(projection reload + reapplication of later events), all events
are applied in a seamless manner (no gap in the sequence of events)

One lesson learned from this is that the first event appended in an new
stream (that is, a fresh, empty, stream) has sequence number 1. There isn't
an "event 0".