EventStreamWrapper: make InitializationAsync testable
for this, an internal (therefore callable from the friend unit test
assembly), static Initialize method is created. Because it is static,
it no longer has access to the fields of EventStreamWrapped: those
are abstracted in an interface, IInitFacade, which is passed to
the method. This interface is implemented by EventStreamWrapper.

It will soon be implemented in a unit tests as a mocked implementation.
The purpose is to define, through unit tests, the complete behaviour of
the EventStreamWrapper implementation.