change behaviour of EventStream.DiscardUpTo: unit tests
This is needed for consistency because, after a call to DiscardUpTo,
Sequence has the value of the event seqnum about to be returned by
TryGetNext (the "next event seqnum"); whereas the rest of the time, after
a call to TryGetNext, Sequence is the seqnum of the event that's just been
returned (the previous event seqnum).

the new expected behaviour is documented in xml doc for the function.
After review, one of the asserts in event_stream.cs also needs be changed.

Also, this breaks the test in wrapper of initialization. These will be
fixed later.