Allow `;Container=` configuration on file paths as well
The main philosophy of connection strings should be that a connection string
maps to an event stream, both for azure streams and for on-disk streams.

However, some applications need to deal with multiple streams found in
the same "location" (where that location can be a parent folder for on-disk
streams, or a Blob Storage account for azure streams), and for those
applications selecting a specific stream should be done in an uniform way.

This commit introduces the notion of "location strings", which can be
file paths (for on-disk streams) or Azure Blob Storage connection strings.
To select a specific stream within a location, it is enough to append
`;Container=<name>` to the location string. This was already the case for
blob storage, so it only extends the on-disk storage:

    C:\LokadData\streams;Container=serviceA
    // Equivalent to:
    C:\LokadData\streams\serviceA