Add support for blob compaction
With Append and Block Blobs, optimal performance can only be achieved
by using 4MB blocks or appends. The nature of event streams implies
that small events are written frequently, resulting in blobs that
are 80x slower to read.

Previously, the only solution was to perform regular copies of the
stream using the "backup" tool, which would perform the copy using
4MB blocks.

This commit introduces a new kind of blob: "compact" blobs, stored in
the same container but as block blobs instead of append blobs. The
naming convention is that `events.00042.compact` is the binary
concatenation of append blobs `events.00000` to `events.00042`
inclusive. When a compact blob is detected, it is used in place of
the range of normal blobs and results in a 80x speedup on a badly
fragmented event stream.

The compact blobs are automatically created once the reader detects
a full append blob (one that has reached the 50000 append limit).
This process is completely transparent to the end user. Obviously,
this creation cannot be performed if the stream is visited with
read-only credentials.

Older versions of Lokad.AzureEventStore do not interact with compact
blobs, and so can work alongside this version on the same event
streams.