Fix bug in stream compaction
A bug was triggered when an application had a high enough uptime to
start two new blobs (at least 50k events) without restarting. The
`AzureStorageDriver` kept a cache containing the starting position and
sequence number of each blob, and this cache would be mangled when the
second new blob was created and triggered a compaction.

The new logic detects when a compaction has occurred, and invalidates
the entire cache (this is not a significant performance problem, since
after a compaction there are rarely more than two blobs).

Bumped to version 1.2.1