Do not wrap exceptions in AppendEventsAsync
When calling AppendEventsAsync in certain configurations (specifically,
those that did not return an `AppendResult<T>` but merely an `AppendResult`,
the code used `ContinueWith` to transform a `Task<AppendResult<T>>` into a
`Task<AppendResult>` (necessary through lack of covariance).

However, doing so would cause any exceptions thrown in the callback passed
to `AppendEventsAsync` to be wrapped in an `AggregateException`.

By using async/await, no wrapping occurs, and the exception leaves `AppendEventsAsync`
exactly as it was thrown inside the callback.