Remove requirement for events to have [DataContract] annotation
The requirement was related to internal Lokad code formatting practices (we do want
the `[DataContract]` annotation to be present) but serves no purpose in a public
library.

Upgrade to 3.0.0 !

This change is not backwards compatible : if a code base contains two non-abstract
classes implementing the event interface, both having the same name, and only one
of which has a `[DataContract]` annotation, then that code base would work
correctly in 2.1.1 but would fail with "Two types named 'Foo' found" in 3.0.0.

Fixes #8