Fix locking issue
The custom implementation of `AsyncQueue` was broken because, inside a `lock`,
it called a function which itself called `TaskCompletionSource<_>.SetResult`
without a `RunContinuationsAsynchronously` setting. This meant that the lock
remained engaged until the thread could process all the synchronous operations
that were waiting for an element to be enqueued, which could take a while (it
could include parsing several MB of event data and applying it through the
projection).

In the presence of many threads performing this operation, it would cause the
thread pool to starve by blocking all threads on the lock.

This design has now been completely dropped, and replaced with a "sleepy process"
pattern. The `EventStreamService` is now a process that can be woken up to
refresh the state by polling the stream (every 30 seconds) and given messages,
which are provided as an inbox to be processed in order. This design is
completely lock-free;