Fix race condition in EventStreamService.CurrentState
In the previous implementation, it was possible for CurrentState to return a
state that did not include very recent changes (the time window was the
duration of a refresh request, or ~10ms).

This commit completely reworks the implementation behind CurrentState: rather
than having requests for CurrentState "discover" that a refresh occurred since
they were requested (which is error-prone), they actively ask for a refresh,
which can only be cleared by starting a new refresh after the request has been
made.