Protect against long-running Azure Blob Storage queries
Since early 2020, the Azure Blob Storage infrastructure has started to
"lose" queries. The symptoms are a request that takes several minutes
(up to 15 minutes observed in some cases), and usually results in failure.
Attempting another request immediately will yield the usual performance
(a few milliseconds).

This commit adds a `AzureHelpers.RetryAsync` method which is used to
wrap read requests that can be delayed in this way. If the request is
observed to take more than 10 seconds, it is aborted and restarted
immediately.