Introduce Lokad.AzureEventStore.Cache package
This introduces **breaking changes** to the way projection cache providers
are implemented in Lokad.AzureEventStore. This corresponds to version 5.0.0 of
the package.

Created new package Lokad.AzureEventStore.Cache that contains cache providers,
currently at version 1.0.0 (and does not need to be included if caching is not
necessary).

`IProjectionCacheProvider.OpenReadAsync` now enumerates all available inputs
instead of just the latest one. This allows better recovery when the most recent
cache file is broken.

`IProjectionCacheProvider.OpenWriteAsync` becomes `TryWriteAsync`. Instead of
returning a stream (and having no control over when that stream is closed
or disposed), it invokes a method on the stream, and can thus trigger an
operation (such as publishing to a blob) after the writing has completed
successfully.

Class `FileProjectionCache` has been moved and renamed to
`Lokad.AzureEventStore.Cache.FileCacheProvider`.

Class `AzureCacheProvider` constructor no longer takes configuration options
(beyond the Blob Storage container) ; these are now specified with writable
properties `AllowWrite`, `AllowRead` and `MaxCacheBlobsCount`.

Added `MappedCacheProvider` which memory-maps an on-disk file instead of just
streaming it in. This allows the use of collections from Lokad.LargeImmutable
to load cached state data from the disk on demand, instead of having to
deserialize it all on startup.

The ExampleProject now includes an example of memory-mapped cache.