Allow saving intermediate caches during initial catchup
Previously, the entire stream had to be reloaded (and passed to the
projection) before `TrySaveAsync` could be invoked to create a
checkpoint. This meant that if the catch-up crashed, all progress
since the last checkpoint was lost.

Moreover, with the introduction of `Lokad.LargeImmutable`, saving to
disk becomes an essential strategy to reduce the memory usage of
the cache (since the cache can then be kept in the disk instead of
in managed memory). Again, doing this during the initial catch-up
was impossible, and so it was necessary for the entire state to
fit in memory before it could be dumped to disk.

This commit introduces `EventStreamService.EventsBetweenCacheSaves`,
by default `uint.MaxValue`. Once this number of events is reached
during a stream catch-up (for large values, this will only be the
initial catch-up), the loading of events stops, the state is then
cycled through `TrySaveAsync` then `TryLoadAsync`, and the loading
resumes. This provides an opportunity to have an intermediate
checkpoint, and to reduce memory usage.