Do not write events that would cause a projection to throw
This commit introduces a pre-write check that invokes all the registered 
projections as part of `AppendEventsAsync()`, discarding the results if successful,
and letting any exception bubble up. 

If an exception is thrown, none of the events are written.

This ensures that the application cannot write an event that will immediately
cause the projection to fail after it is read back.