Track append blobs in a compacted block blob
When restarting an application, it loads from the cache the state after
event N, then it finds the event N in the event stream and starts
reading from there. To find the event N in the event stream, it
determines which of the blobs in the stream contain that event (the
start event of each blob is easily known), and then reads through that
blob until event N is reached.

This used to be fast, because each append blob contains up to 50k writes
(as an Azure Blob Storage constraint), so it’s fast to stream through it.

However, since we introduced compaction, the stream is instead composed
of only two blobs: the last append blob (not yet reached 50k writes),
and a block blob that contains all events before that. If the
application restarts and looks for an event N that is not in the last
append blob, it will need to read through the entire compacted block blob
(which may contain several gigabytes of data).

This commit introduces the `EventBlob` abstraction, which corresponds
to an individual append blob in the stream, and is thus able to track
the sequence number of its first event. Reading from an `EventBlob`
will read from the correct position of the compacted blob, if present,
and otherwise from the append blob.