Projections: Allow to restore an external state and to commit additional changes to the projection.
`ReifiedProjection` now exits the constructor with a null current state and must be initialized by calling `CreateAsync`.
A `StorageProvider` is provided to the EventStreamService to manage the state external storage.
`IProjection.TryRestoreAsync` tries to load an external state.
`IProjection.CommitAsync` marks ‘state’ as being the latest in the sequence of states produced by applying events persisted in the stream (as opposed to tentative state instances that are produced by applying tentative events that will not be persisted). This gives the projection the liberty to perform any operations related to the persistence of the state, such as flushing parts of it to an external state that may be loaded later.