Use Memory<byte> instead of BinaryReader for events
The RawEvent now keeps a reference to existing memory instead 
of having its own internal byte[]. 

The IStorageDriver.ReadAsync now takes the backing which will 
be used for all returned RawEvent instances. 

The callback returned by FetchBackgroundAsync now does 
nothing and returns true if the EventStream._cache 
contains unconsumed events (this is because EventStream 
now uses its own memory buffers, instead of those being 
handled by the driver).

As a benefit, we can now detect that the end of the current
page has been reached by looking at the bytes, instead of
waiting for an end-of-stream exception to be thrown.